projects/multi-agent-system/requirements.txt
# 多Agent协作系统 - 依赖包

# 核心框架
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# HTTP客户端
httpx>=0.25.0

# 异步支持
asyncio-pool>=0.6.0

# 数据库
redis>=5.0.0
chromadb>=0.4.0

# 工具
python-multipart>=0.0.6
orjson>=3.9.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# 可选：前端界面
# gradio>=4.0.0
# streamlit>=1.28.0
